refactor: merge async APIs into main modules
